requirements.txt
speechrecognition
pyttsx3
openai
python-dotenv
